feat: ubtc extension
