+++ typescript/infra/UBTC/bsquared-strategy.yaml
@@ -1,0 +1,30 @@
+boba:
+  submitter:
+    chain: boba
+    safeAddress: "0x207FfFa7325fC5d0362aB01605D84B268b61888f"
+    type: gnosisSafe
+    version: "1.0"
+bsquared:
+  submitter:
+    chain: bsquared
+    safeAddress: "0x7A363efD42305BeDBA307d25351F8ea157b69A1A"
+    type: gnosisSafe
+    version: "1.0"
+nibiru:
+  submitter:
+    chain: nibiru
+    type: jsonRpc
+    version: "1.0"
+soneium:
+  submitter:
+    chain: soneium
+    safeAddress: "0x8433e6e9183B5AAdaf4b52c624B963D95956e3C9"
+    type: gnosisSafe
+    version: "1.0"
+swell:
+  submitter:
+    chain: swell
+    safeAddress: "0xC11e22A31787394950B31e2DEb1d2b5546689B65"
+    type: gnosisSafe
+    version: "1.0"
+
